% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{add_object_tier}
\alias{add_object_tier}
\title{Adds an element named <obj_type> to content sub-element.}
\usage{
add_object_tier(obj, obj_type)
}
\arguments{
\item{obj}{A list.}

\item{obj_type}{A non-empty string.}
}
\value{
A list containing the \code{<obj>$content$<obj_type>}, whose value is
  either \code{NULL} if it had not any value before or the value of
  \code{<obj>$content}.
}
\description{
This function takes a list, sees if it has an element named 'content' and
then checks if content contains an element named \code{obj_type}: if it does
not contain an element as passed in \code{obj_type} this function adds it,
otherwise it leaves the list \code{obj} untouched. The contents of 'content'
are set as value of \code{<obj>$content$<obj_type>}.
}
\keyword{internal}
