% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_quotes}
\alias{add_quotes}
\title{Wrap a string in quotes.}
\usage{
add_quotes(str)
}
\arguments{
\item{str}{A character vector of strings.}
}
\value{
A vector of string wrapped in quotes.
}
\description{
Wrap a string in quotes.
}
\examples{

add_quotes("a") # "a"

}
\keyword{internal}
