% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{allele_name}
\alias{allele_name}
\title{Extract allele names from strings of the form rs123-G}
\usage{
allele_name(risk_allele_names)
}
\arguments{
\item{risk_allele_names}{}
}
\value{
A character vector of allele names.
}
\description{
This function parses strings of the form \code{"rs123-G"} and returns
the name of the allele; it uses the regex \code{([ATCG]+)$}.
}
\keyword{internal}
