% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{ancestral_groups_tbl}
\alias{ancestral_groups_tbl}
\title{Creates an ancestral groups table.}
\usage{
ancestral_groups_tbl(
  study_id = character(),
  ancestry_id = integer(),
  ancestral_group = character()
)
}
\arguments{
\item{study_id}{GWAS Catalog study accession identifier.}

\item{ancestry_id}{Ancestry identifier.}

\item{ancestral_group}{Ancestral group.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a ancestral groups table.
}
\keyword{internal}
