% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{ancestries_tbl}
\alias{ancestries_tbl}
\title{Creates an ancestries table.}
\usage{
ancestries_tbl(
  study_id = character(),
  ancestry_id = integer(),
  type = character(),
  number_of_individuals = integer()
)
}
\arguments{
\item{study_id}{GWAS Catalog study accession identifier.}

\item{ancestry_id}{Ancestry identifier.}

\item{type}{Type of cohort sample, either \code{"initial"} or
\code{"replication"}.}

\item{number_of_individuals}{Number of individuals in the cohort sample.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an ancestries table.
}
\keyword{internal}
