% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{association_to_trait}
\alias{association_to_trait}
\title{Map an association id to an EFO trait id}
\usage{
association_to_trait(association_id, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{association_id}{A character vector of association accession
identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}
}
\value{
A dataframe of two identifiers. First column is the association
  identifier and the second column is the EFO trait identifier.
}
\description{
Map an association accession identifier to an EFO trait id.
}
\examples{
# Map GWAS association identifiers to EFO trait identifiers
association_to_trait(c('24300097', '24299759'))

}
