% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{association_to_variant}
\alias{association_to_variant}
\title{Map an association id to a variant id}
\usage{
association_to_variant(association_id, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{association_id}{A character vector of association accession
identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}
}
\value{
A dataframe of two identifiers. First column is the association
  identifier and the second column is the variant identifier.
}
\description{
Map an association accession identifier to a variant identifier.
}
\examples{
# Map GWAS association identifiers to variant identifiers
association_to_variant(c('24300097', '24299759'))

}
