% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\docType{class}
\name{associations-class}
\alias{associations-class}
\title{An S4 class to represent a set of GWAS Catalog associations}
\description{
The association object consists of four slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of GWAS Catalog associations. Each association is an observation (row) in the
associations table --- main table. All tables have the column \code{association_id} as
primary key.
}
\section{Slots}{

\describe{
\item{\code{associations}}{A \code{\link[tibble]{tibble}} listing
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{pvalue}{TODO}
\item{pvalue_description}{TODO}
\item{pvalue_mantissa}{TODO}
\item{pvalue_exponent}{TODO}
\item{multiple_snp_haplotype}{TODO}
\item{snp_interaction}{TODO}
\item{snp_type}{TODO}
\item{standard_error}{TODO}
\item{range}{TODO}
\item{or_per_copy_number}{TODO}
\item{beta_number}{TODO}
\item{beta_unit}{TODO}
\item{beta_direction}{TODO}
\item{beta_description}{TODO}
\item{last_mapping_date}{TODO}
\item{last_update_date}{TODO}
}}

\item{\code{loci}}{A \code{\link[tibble]{tibble}} listing loci. Columns:
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{locus_id}{TODO}
\item{haplotype_snp_count}{TODO}
\item{description}{TODO}
}}

\item{\code{risk_alleles}}{A \code{\link[tibble]{tibble}} listing risk alleles. Columns:
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{locus_id}{TODO}
\item{variant_id}{TODO}
\item{risk_allele}{TODO}
\item{risk_frequency}{TODO}
\item{genome_wide}{TODO}
\item{limited_list}{TODO}
}}

\item{\code{genes}}{A \code{\link[tibble]{tibble}} listing author reported genes. Columns:
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{locus_id}{TODO.}
\item{gene_name}{TODO.}
}}

\item{\code{ensembl_ids}}{A \code{\link[tibble]{tibble}} listing Ensembl gene identifiers. Columns:
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{locus_id}{TODO.}
\item{gene_name}{Gene symbol according to \href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}
\item{entrez_id}{The Entrez identifier of a gene, see ref.
\href{https://dx.doi.org/10.1093\%2Fnar\%2Fgkq1237}{10.1093/nar/gkq1237} for
more information.}
\item{ensembl_id}{The Ensembl identifier of an Ensembl gene, see Section
\href{https://www.ensembl.org/info/genome/genebuild/genome_annotation.html}{Gene
annotation in Ensembl} for more information.}
}}

\item{\code{entrez_ids}}{A \code{\link[tibble]{tibble}} listing Entrez gene identifiers. Columns:
\describe{
\item{association_id}{GWAS Catalog association accession identifier, e.g., \code{20250}.}
\item{locus_id}{TODO.}
\item{gene_name}{Gene symbol according to \href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}
\item{entrez_id}{The Entrez identifier of a gene, see ref.
\href{https://dx.doi.org/10.1093\%2Fnar\%2Fgkq1237}{10.1093/nar/gkq1237} for
more information.}
}}
}}

