% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{associations}
\alias{associations}
\title{Constructor for the S4 associations object.}
\usage{
associations(associations = associations_tbl(), loci = loci_tbl(),
  risk_alleles = risk_alleles_tbl(), genes = reported_genes_tbl(),
  ensembl_ids = ensembl_ids_tbl(), entrez_ids = entrez_ids_tbl())
}
\arguments{
\item{associations}{TODO.}

\item{loci}{TODO.}

\item{risk_alleles}{TODO.}

\item{genes}{TODO.}

\item{ensembl_ids}{TODO.}

\item{entrez_ids}{TODO.}
}
\value{
An object of class \linkS4class{associations}.
}
\description{
Constructor for the S4 \linkS4class{associations} object.
}
\keyword{internal}
