% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{associations_drop_na}
\alias{associations_drop_na}
\title{Drop any NA associations.}
\usage{
associations_drop_na(s4_associations)
}
\arguments{
\item{s4_associations}{An object of class \linkS4class{associations}.}
}
\value{
An object of class \linkS4class{associations}.
}
\description{
This function takes an associations S4 object and removes any association
identifiers that might have been NA. This ensures that there is always a
non-NA \code{association_id} value in all tables. This is important as the
\code{association_id} is the primary key.
}
\keyword{internal}
