% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{associations_tbl}
\alias{associations_tbl}
\title{Creates an associations table.}
\usage{
associations_tbl(association_id = integer(), pvalue = double(),
  pvalue_description = character(), pvalue_mantissa = integer(),
  pvalue_exponent = integer(), multiple_snp_haplotype = logical(),
  snp_interaction = logical(), snp_type = character(),
  standard_error = double(), range = character(),
  or_per_copy_number = double(), beta_number = double(),
  beta_unit = character(), beta_direction = character(),
  beta_description = character(),
  last_mapping_date = lubridate::ymd_hms(),
  last_update_date = lubridate::ymd_hms())
}
\arguments{
\item{association_id}{TODO.}

\item{pvalue}{TODO.}

\item{pvalue_description}{TODO.}

\item{pvalue_mantissa}{TODO.}

\item{pvalue_exponent}{TODO.}

\item{multiple_snp_haplotype}{TODO.}

\item{snp_interaction}{TODO.}

\item{snp_type}{TODO.}

\item{standard_error}{TODO.}

\item{range}{TODO.}

\item{or_per_copy_number}{TODO.}

\item{beta_number}{TODO.}

\item{beta_unit}{TODO.}

\item{beta_direction}{TODO.}

\item{beta_description}{TODO.}

\item{last_mapping_date}{TODO.}

\item{last_update_date}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an associations table.
}
\keyword{internal}
