% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{contains_question_mark}
\alias{contains_question_mark}
\title{Does a string contain a question mark?}
\usage{
contains_question_mark(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings contain a question mark. This function uses the following
regular expression: \code{[\\?]}.
}
\keyword{internal}
