% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{cytogenetic_band_to_genomic_range}
\alias{cytogenetic_band_to_genomic_range}
\title{Convert a cytogenetic band string to genomic coordinates.}
\usage{
cytogenetic_band_to_genomic_range(bands)
}
\arguments{
\item{bands}{A \code{\link[base]{character}} vector of cytogenetic bands.}
}
\value{
A dataframe of genomic ranges. Columns are: chromosome, start and
  end. Each row corresponds to a queried cytogenetic band (in the same order
  as queried).
}
\description{
This function uses the provided \code{\link[gwasrapidd]{cytogenetic_bands}}
dataframe to convert cytogenetic band names to genomic coordinates.
}
\keyword{internal}
