% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_links}
\alias{drop_links}
\title{Discard elements whose name starts with "_links".}
\usage{
drop_links(lst)
}
\arguments{
\item{lst}{A list (remember that a dataframe is also a list).}
}
\value{
A list (dataframe) with potentially less elements (columns if dataframe).
}
\description{
Takes a list and discards any elements whose name starts with
\code{"_links"}. \code{drop_links} can be applied on column names of
dataframes because dataframes are also lists after all.
}
