% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Converts an empty vector to a scalar NA}
\usage{
empty_to_na(x)
}
\arguments{
\item{x}{An atomic vector, of one of these types: \code{character},
\code{integer}, \code{double} or \code{logical}.}
}
\value{
An atomic vector of the same type as \code{x}. If \code{x} is empty
  then \code{NA} is returned, otherwise \code{x} is returned as is.
}
\description{
This function converts an empty vector to a scalar NA of the same type as the
input.
}
\keyword{internal}
