% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{ensembl_ids_tbl}
\alias{ensembl_ids_tbl}
\title{Creates an Ensembl gene identifiers' table.}
\usage{
ensembl_ids_tbl(association_id = integer(), locus_id = integer(),
  gene_name = character(), ensembl_id = character())
}
\arguments{
\item{association_id}{TODO.}

\item{locus_id}{TODO.}

\item{gene_name}{TODO.}

\item{ensembl_id}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an Ensembl gene identifiers' table.
}
\keyword{internal}
