% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{extract_association_id}
\alias{extract_association_id}
\title{Extract association identifiers from URLs}
\usage{
extract_association_id(urls)
}
\arguments{
\item{urls}{A character vector of URLs of the form
\code{".*{association_id}$"}.}
}
\value{
A character vector of association identifiers.
}
\description{
This function extracts association identifiers from URLs of the form:
\code{".*{association_id}$"}.
}
\keyword{internal}
