% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Filter GWAS Catalog objects by identifier.}
\usage{
filter_by_id(x, id)
}
\arguments{
\item{x}{An object of class either \linkS4class{studies},
\linkS4class{associations}, \linkS4class{variants}, or
\linkS4class{traits}.}

\item{id}{Identifier.}
}
\value{
Returns an object of class either \linkS4class{studies},
  \linkS4class{associations}, \linkS4class{variants}, or
  \linkS4class{traits}.
}
\description{
Use \code{filter_by_id} to filter GWAS Catalog objects by their respective
identifier (\code{id}).
}
\examples{
# Filter a studies object by identifier.
filter_by_id(studies_ex01, 'GCST001585')

# Filter a associations object by identifier.
filter_by_id(associations_ex01, c('22509', '19537565'))

# Filter a variants object by identifier.
filter_by_id(variants_ex01, 'rs56261590')

# Filter a traits object by identifier.
filter_by_id(traits_ex01, 'EFO_0004884')

}
\keyword{internal}
