% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{filter_genomic_location_by_chr_name}
\alias{filter_genomic_location_by_chr_name}
\title{Filter a genomic location dataframe by chromosome name.}
\usage{
filter_genomic_location_by_chr_name(df, chr_names_to_keep = chromosomes,
  warnings = TRUE)
}
\arguments{
\item{df}{A dataframe with the columns:
\itemize{
\item \code{chromosomeName}
\item \code{chromosomePosition}
\item \code{region.name}
\item \code{_links.snps.href}
}}

\item{chr_names_to_keep}{A character vector of human chromosome names.}

\item{warnings}{Whether to show warnings.}
}
\value{
A dataframe (a \code{\link[tibble]{tibble}} actually) with one row
  only, corresponding to a single genomic location.
}
\description{
Filter genomic location cases by chromosome name. This is useful to filter
out loci mapped to non-standard chromosomes, such as: \itemize{ \item GRC
assembly patches \item haplotype (HAPs) \item pseudo autosomal regions (PARs)
}
Only genomic locations lying in \code{chr_names_to_keep} are kept, and then, if more
than one is still present, we return only one, the first.
}
\keyword{internal}
