% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gc_examples}
\alias{gc_examples}
\alias{studies_ex01}
\alias{studies_ex02}
\alias{associations_ex01}
\alias{associations_ex02}
\alias{variants_ex01}
\alias{variants_ex02}
\alias{traits_ex01}
\alias{traits_ex02}
\title{gwasrapidd entities' examples}
\format{
\describe{
  \item{studies_ex01}{An S4 \linkS4class{studies} object of 2 studies:
  \code{'GCST001585'} and \code{'GCST003985'}.}
  \item{studies_ex02}{An S4 \linkS4class{studies} object of 2 studies:
  \code{'GCST001585'} and \code{'GCST006655'}.}
  \item{associations_ex01}{An S4 \linkS4class{associations} object of 4
  associations: \code{'22509'}, \code{'22505'}, \code{'19537565'} and
  \code{'19537593'}.}
  \item{associations_ex02}{An S4 \linkS4class{associations} object of 3
  associations: \code{'19537593'}, \code{'31665940'} and \code{'34944736'}.}
  \item{variants_ex01}{An S4 \linkS4class{variants} object of 3 variants:
  \code{'rs146992477'}, \code{'rs56261590'} and \code{'rs4725504'}.}
  \item{variants_ex02}{An S4 \linkS4class{variants} object of 4 variants:
  \code{'rs56261590'}, \code{'rs4725504'}, \code{'rs11099757'} and
  \code{'rs16871509'}.}
  \item{traits_ex01}{An S4 \linkS4class{traits} object of 3 traits:
  \code{'EFO_0004884'}, \code{'EFO_0004343'} and \code{'EFO_0005299'}.}
  \item{traits_ex02}{An S4 \linkS4class{traits} object of 4 traits:
  \code{'EFO_0007845'}, \code{'EFO_0004699'}, \code{'EFO_0004884'} and
  \code{'EFO_0004875'}.}
}

An object of class \code{studies} of length 1.

An object of class \code{associations} of length 1.

An object of class \code{associations} of length 1.

An object of class \code{variants} of length 1.

An object of class \code{variants} of length 1.

An object of class \code{traits} of length 1.

An object of class \code{traits} of length 1.
}
\usage{
studies_ex01

studies_ex02

associations_ex01

associations_ex02

variants_ex01

variants_ex02

traits_ex01

traits_ex02
}
\description{
These are examples of GWAS Catalog entities shipped with gwasrapidd:
}
\keyword{datasets}
