% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{genomic_contexts_tbl}
\alias{genomic_contexts_tbl}
\title{Creates a genomic contexts table.}
\usage{
genomic_contexts_tbl(variant_id = character(), gene_name = character(),
  chromosome_name = character(), chromosome_position = integer(),
  distance = integer(), is_closest_gene = logical(),
  is_intergenic = logical(), is_upstream = logical(),
  is_downstream = logical(), source = character(),
  mapping_method = character())
}
\arguments{
\item{variant_id}{TODO.}

\item{gene_name}{TODO.}

\item{chromosome_name}{TODO.}

\item{distance}{TODO.}

\item{is_closest_gene}{TODO.}

\item{is_intergenic}{TODO.}

\item{is_upstream}{TODO.}

\item{is_downstream}{TODO.}

\item{source}{TODO.}

\item{mapping_method}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a genomic contexts table.
}
\keyword{internal}
