% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{genotyping_techs_tbl}
\alias{genotyping_techs_tbl}
\title{Creates a genotyping technologies table.}
\usage{
genotyping_techs_tbl(
  study_id = character(),
  genotyping_technology = character()
)
}
\arguments{
\item{study_id}{GWAS Catalog study accession identifier.}

\item{genotyping_technology}{Genotyping technology.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a genotyping technologies table.
}
\keyword{internal}
