% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{GWAS Catalog Metadata}
\usage{
get_metadata(warnings = TRUE)
}
\arguments{
\item{warnings}{A logical value indicating whether to trigger a warning if the request is not successful.}
}
\value{
A named character vector whose names are:
}
\description{
Provides a list of the resources the GWAS Catalog data is currently mapped
against, namely:
\itemize{
\item Ensembl release number;
\item Genome build version;
\item dbSNP version.
}
}
\details{
In addition, the date since this combination of resource versions has been
in use is also returned.
}
\examples{
\donttest{get_metadata()
# ensembl_release_number  genome_build_version  dbsnp_version                 usage_start_date
#                   "94"          "GRCh38.p12"          "151"   "2018-10-02T13:00:02.054+0000"
}

}
