% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{get_snps}
\alias{get_snps}
\title{Get SNPs by rsId.}
\usage{
get_snps(snp_ids, verbose = FALSE, warnings = TRUE,
  remove_duplicated_snps = FALSE)
}
\arguments{
\item{snp_ids}{SNP Ids, starting with "rs".}

\item{verbose}{Whether to be chatty or not.}

\item{warnings}{Whether to print warnings or not.}

\item{collapse}{Whether to bind all queries into one single dataframe.}
}
\value{
TODO.
}
\description{
Retrieve SNPs by Id.
}
\examples{
get_snps(c("rs6431996", "rs7546498", "rs123"))

}
