% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{get_snps_by_id}
\alias{get_snps_by_id}
\title{Get SNPs by rsID.}
\usage{
get_snps_by_id(snp_ids, verbose = FALSE, warnings = TRUE,
  remove_duplicated_snps = FALSE)
}
\arguments{
\item{snp_ids}{SNP IDs, starting with "rs".}

\item{verbose}{Whether to be chatty or not.}

\item{warnings}{Whether to print warnings or not.}

\item{remove_duplicated_snps}{Whether to remove duplicated SNPs.}
}
\value{
A \code{\link[tibble]{tibble}} where rows are SNPs and columns are
  the following variables:
  \itemize{
  \item \code{rsId}, Reference SNP cluster ID (rsID);
  \item \code{merged}, whether this SNP has been merged with another SNP in a newer genome build;
  \item \code{chromosomeName}, the name of the chromosome where the SNP is located;
  \item \code{chromosomePosition}, the genomic position along the chromosome (1-based indexing);
  \item \code{region.name}, the genomic location of this SNP according to the
  \href{https://en.wikipedia.org/wiki/Locus_(genetics)}{cytogenetic banding
  nomenclature}; \item \code{functionalClass}, the functional class of the SNP;
  \item \code{lastUpdateDate}, the last date this mapping information
  was updated;
  \item \code{genomicContexts}, the genomic contexts for this
  SNP, including upstream, downstream and mapped genes. This is a
  \code{\link[tibble]{tibble}} of genomic contexts for each SNP, containing the columns:
  \itemize{
  \item \code{isIntergenic} TODO.
  \item \code{isUpstream} TODO.
  \item \code{isDownstream} TODO.
  \item \code{distance} TODO.
  \item \code{source} TODO.
  \item \code{mappingMethod} TODO.
  \item \code{isClosestGene} TODO.
  \item \code{gene.geneName} TODO.
  \item \code{gene.entrezGeneIds} TODO.
  \item \code{gene.ensemblGeneIds} TODO.
  \item \code{location.chromosomeName} TODO.
  \item \code{location.chromosomePosition} TODO.
  \item \code{location.region.name} TODO.
  \item \code{location._links.snps.href} TODO.
  }
}
}
\description{
Retrieve SNPs by Reference SNP cluster ID (rsID).
}
\examples{
\donttest{get_snps_by_id(c("rs6431996", "rs7546498", "rs123"))}

}
