% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{get_snps_by_location}
\alias{get_snps_by_location}
\title{Retrieve SNPs by genomic location.}
\usage{
get_snps_by_location(chr, start, end, verbose = FALSE,
  warnings = FALSE, drop_links = TRUE, query_grange = TRUE,
  remove_duplicated_snps = FALSE)
}
\arguments{
\item{chr}{A character vector of human chromosome names.}

\item{start}{An integer vector of start positions.}

\item{end}{An integer vector of end positions.}

\item{verbose}{Whether to be chatty or not.}

\item{warnings}{Whether to print warnings or not.}

\item{drop_links}{Whether to drop columns related with links to resources.}

\item{query_grange}{Whether to keep an extra column of the genomic range
query.}

\item{remove_duplicated_snps}{Whether to remove duplicated SNPs in output.
This can happen when querying several genomic intervals that overlap with
each other.}
}
\value{
A \code{\link[tibble]{tibble}} where rows are SNPs and columns are:
  \itemize{
  \item \code{query_grange}, the query genomic range (if
  \code{query_grange == TRUE});
  \item \code{rsId}, the SNP Id;
  \item \code{merged}, Whether this SNP has been merged with another SNP in a
  newer genome build;
  \item \code{chromosomeName}, the name of the chromosome where the SNP is
  located;
  \item \code{chromosomePosition}, the genomic position
  along the chromosome (1-based indexing); \item \code{region.name}, the
  genomic location of this SNP according to the
  \href{https://en.wikipedia.org/wiki/Locus_(genetics)}{cytogenetic banding
  nomenclature}; \item \code{functionalClass}, the SNP’s functional class;
  \item \code{lastUpdateDate}, the last date this SNP’s mapping information
  was updated;
  \item \code{genomicContexts}, the genomic contexts for this
  SNP, including upstream, downstream and mapped genes. This is a
  \code{\link[tibble]{tibble}} of genomic contexts for each SNP, containing the columns:
  \itemize{
  \item \code{isIntergenic} TODO.
  \item \code{isUpstream} TODO.
  \item \code{isDownstream} TODO.
  \item \code{distance} TODO.
  \item \code{source} TODO.
  \item \code{mappingMethod} TODO.
  \item \code{isClosestGene} TODO.
  \item \code{gene.geneName} TODO.
  \item \code{gene.entrezGeneIds} TODO.
  \item \code{gene.ensemblGeneIds} TODO.
  \item \code{location.chromosomeName} TODO.
  \item \code{location.chromosomePosition} TODO.
  \item \code{location.region.name} TODO.
  \item \code{location._links.snps.href} TODO.
  }
}
}
\description{
Retrieve SNPs by genomic location. Genomic ranges are provided
through the three arguments: \code{chr}, \code{start}, and \code{end}. Genomic ranges
are assembled from these three vectors by matching elements by position. Thus,
all three vectors must have the same length.
}
\examples{
\donttest{get_snps_by_location("X", 1L, 20000L)
get_snps_by_location("6", 1L, 250000L)}

}
