% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{get_studies_by_id}
\alias{get_studies_by_id}
\title{Get association study by accession ID}
\usage{
get_studies_by_id(study_ids, verbose = FALSE, warnings = TRUE,
  remove_duplicated_studies = FALSE)
}
\arguments{
\item{study_ids}{study IDs as a character vector, e.g., "GCST000854".}

\item{verbose}{Whether to be chatty or not.}

\item{warnings}{Whether to print warnings or not.}

\item{remove_duplicated_studies}{Whether to remove duplicated studies.}
}
\value{
A \code{\link[tibble]{tibble}} where rows are studies and columns are
  the following variables:
  \itemize{
  \item \code{accessionId}, GWAS Catalog study accession ID;
  \item \code{diseaseTrait_trait}, Free text description of the trait investigated in this study;
  \item \code{initialSampleSize}, 	Initial sample description;
  \item \code{replicationSampleSize}, Replication sample description;
  \item \code{gxe}, Whether the study investigates a gene-environment interaction;
  \item \code{gxg}, Whether the study investigates a gene-gene interaction;
  \item \code{snpCount}, Number of SNPs passing QC;
  \item \code{qualifier}, Qualifier of number of SNPs passing QC (e.g. >);
  \item \code{imputed}, Whether SNPs were imputed;
  \item \code{pooled}, Whether samples were pooled;
  \item \code{studyDesignComment}, Any other relevant study design information;
  \item \code{fullPvalueSet}, Whether full summary statistics are available for this study;
  \item \code{userRequested}, Whether the addition of this study to the Catalog was requested by a user;
  \item \code{ancestries}, Ancestry entries for this study. A list of
  \code{\link[tibble]{tibble}}s whose columns are:
  \itemize{
  \item \code{type}, TODO;
  \item \code{numberOfIndividuals}, TODO;
  \item \code{ancestralGroups}, TODO;
  \item \code{countryOfOrigin}, TODO;
  \item \code{countryOfRecruitment}, TODO;
  }
  \item \code{genotypingTechnologies}, Genotyping technology used in this study;
  \item \code{publicationInfo}, Convenience representation of the publication.
  A list of #'   \code{\link[tibble]{tibble}}s whose columns are:
  \itemize{
  \item \code{pubmedId}, PubMed ID;
  \item \code{publicationDate}, Publication date;
  \item \code{publication}, Journal;
  \item \code{title}, Title of the association study;
  \item \code{author_fullname}, First author name;
  }
  \item \code{platforms_manufacturer}, Genotyping platform(s) used in this study;
}
}
\description{
Get association studies in the GWAS Catalog by their
accession ID.
}
\examples{
\donttest{get_studies_by_id(c("GCST000854", "GCST005268"))}

}
