% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_studies.R
\name{get_studies_by_pubmed_id}
\alias{get_studies_by_pubmed_id}
\title{Get GWAS Catalog studies by PubMed identifiers}
\usage{
get_studies_by_pubmed_id(
  pubmed_id = NULL,
  verbose = FALSE,
  warnings = TRUE,
  page_size = 20L
)
}
\arguments{
\item{pubmed_id}{An integer vector of
\href{https://pubmed.ncbi.nlm.nih.gov/}{PubMed} identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}

\item{page_size}{An integer scalar indicating the
\href{https://www.ebi.ac.uk/gwas/rest/docs/api#_paging_resources}{page}
value to be used in the JSON requests, can be between \code{1} and
\code{1000}.}
}
\value{
A \linkS4class{studies} object.
}
\description{
Gets studies whose associated publications match
 \href{https://pubmed.ncbi.nlm.nih.gov/}{PubMed} identifiers.
}
\keyword{internal}
