% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variants.R
\name{get_variants}
\alias{get_variants}
\title{Get GWAS Catalog variants}
\usage{
get_variants(study_id = NULL, association_id = NULL,
  variant_id = NULL, efo_id = NULL, pubmed_id = NULL,
  genomic_range = NULL, gene_name = NULL, efo_trait = NULL,
  reported_trait = NULL, set_operation = "union", verbose = FALSE,
  warnings = TRUE)
}
\arguments{
\item{study_id}{A \code{character} vector of GWAS Catalog study accession
identifiers.}

\item{association_id}{An \code{integer} vector of GWAS Catalog association
identifiers.}

\item{variant_id}{A \code{character} vector of GWAS Catalog variant identifiers.}

\item{efo_id}{A character vector of \href{https://www.ebi.ac.uk/efo/}{EFO}
identifiers.}

\item{pubmed_id}{An \code{integer} vector of
\href{https://en.wikipedia.org/wiki/PubMed}{PubMed} identifiers.}

\item{genomic_range}{A named \code{list} of three vectors:
\describe{ \item{chromosome}{A character vector of chromosome names of the
form 1--22, X or Y.} \item{start}{A numeric vector of start positions,
starting at 1.} \item{end}{A numeric vector of end positions.} }
The three vectors need to be of the same length so that \code{chromosome}
names, \code{start} and \code{end} positions are matched by position hence
defining a genomic range.}

\item{gene_name}{Gene symbol according to
\href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}

\item{efo_trait}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} trait descriptions, e.g.,
\code{'uric acid measurement'}.}

\item{reported_trait}{A \code{character} vector of phenotypic traits as
reported by the original authors' the study.}

\item{set_operation}{Either \code{'union'} or \code{'intersection'}. This
tells how variants retrieved by different criteria  should be combined:
\code{'union'} binds together all results removing duplicates and
\code{'intersection'} only keeps same variants found with different
criteria.}

\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if any.}
}
\value{
A \linkS4class{variants} object.
}
\description{
Retrieves variants via the NHGRI-EBI GWAS Catalog REST API. The REST
API is queried multiple times with the criteria passed as arguments (see
below). By default all variants that match the criteria supplied in the
arguments are retrieved: this corresponds to the default option
\code{set_operation} set to \code{'union'}. If you rather have only the
variants that match simultaneously all criteria provided, then set
\code{set_operation} to \code{'intersection'}.
}
