% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variants.R
\name{get_variants_by_gene_name}
\alias{get_variants_by_gene_name}
\title{Get GWAS Catalog variants by gene name.}
\usage{
get_variants_by_gene_name(gene_name = NULL, verbose = FALSE,
  warnings = TRUE, page_size = 1000L)
}
\arguments{
\item{gene_name}{A \code{character} vector of gene names.}

\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if any.}

\item{page_size}{An \code{integer} scalar indicating the
\href{https://www.ebi.ac.uk/gwas/rest/docs/api#_paging_resources}{page}
value to be used in the JSON requests, can be between \code{1} and
\code{1000}.}
}
\value{
A \linkS4class{variants} object.
}
\description{
Gets variants whose genomic context includes a specific gene or genes.
}
\keyword{internal}
