% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variants.R
\name{get_variants_by_reported_trait}
\alias{get_variants_by_reported_trait}
\title{Get GWAS Catalog variants by reported traits}
\usage{
get_variants_by_reported_trait(reported_trait = NULL, verbose = FALSE,
  warnings = TRUE, page_size = 20L)
}
\arguments{
\item{reported_trait}{A character vector of phenotypic traits as
reported by the original authors' the study.}

\item{verbose}{Whether the function should be verbose about the different
queries or not.}

\item{warnings}{Whether to print warnings.}

\item{page_size}{An integer scalar indicating the
\href{https://www.ebi.ac.uk/gwas/rest/docs/api#_paging_resources}{page}
value to be used in the JSON requests, can be between \code{1} and
\code{1000}.}
}
\value{
A \linkS4class{variants} object.
}
\description{
Gets variants that match the reported traits, as reported by the original
authors' of the study.
}
\details{
Note: Right now there is a server-side bug and queries by
\code{reported_trait} are currently case sensitive but this should not be the
case.
}
\keyword{internal}
