% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasrapidd-package.R
\docType{package}
\name{gwasrapidd-package}
\alias{gwasrapidd}
\alias{gwasrapidd-package}
\title{gwasrapidd: REST API Client for the NHGRI-EBI GWAS Catalog}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

gwasrapidd: GWAS R API Data Download.
    gwasrapidd provides easy access to the NHGRI-EBI GWAS Catalog data by
    accessing the REST API https://www.ebi.ac.uk/gwas/docs/api.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/gwasrapidd}
  \item Report bugs at \url{https://github.com/ramiromagno/gwasrapidd/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (0000-0001-5226-3441)

Other contributors:
\itemize{
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (0000-0002-0454-9207) [thesis advisor]
  \item University of Algarve [copyright holder]
  \item Fundação para a Ciência e Tecnologia [funder]
}

}
\keyword{internal}
