% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{is_association_id}
\alias{is_association_id}
\title{Is a string a GWAS Catalog association accession ID?}
\usage{
is_association_id(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid GWAS Catalog association IDs (returns
\code{TRUE}). Association IDs are tested against the following regular
expression: \code{^\\\\d+$}.
}
\keyword{internal}
