% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is_content_empty}
\alias{is_content_empty}
\title{Did the response to a GWAS Catalog REST API request contain no elements?}
\usage{
is_content_empty(response)
}
\arguments{
\item{response}{A response (\code{list}) as returned by \code{\link{request}}.}
}
\value{
A logical value.
}
\description{
Checks if the response contained no elements, i.e., if the content is empty.
}
