% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{is_efo_id}
\alias{is_efo_id}
\title{Is a string an EFO trait ID?}
\usage{
is_efo_id(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid EFO trait IDs (returns
\code{TRUE}). EFO trait IDs are tested against the following regular
expression: \code{^EFO_\\\\d\{7\}$}.
}
\keyword{internal}
