% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is_embedded}
\alias{is_embedded}
\title{Is the GWAS response wrapped in an '_embedded' object?}
\usage{
is_embedded(obj)
}
\arguments{
\item{obj}{The response object as return by \code{jsonlite::fromJSON}.}
}
\value{
A logical value.
}
\description{
Checks if the response is wrapped in an '_embedded' object by checking if an
element named '_embedded' exists.
}
\keyword{internal}
