% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{is_empty_str}
\alias{is_empty_str}
\title{Is string empty or an all whitespace string?}
\usage{
is_empty_str(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A \code{\link[base]{character}} vector.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A \code{\link[base]{logical}}.
}
\description{
Matches the \code{string} vector against \code{"^\\\s*$"}.
}
\keyword{internal}
