% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{is_genomic_range}
\alias{is_genomic_range}
\title{Is a string a valid human genomic range?}
\usage{
is_genomic_range(str, convert_NA_to_FALSE = FALSE)
}
\arguments{
\item{str}{A character vector of putative genomic ranges of the form
\{chr\}:\{start\}-\{end\}.}

\item{convert_NA_to_FALSE}{Whether to preserve \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
Returns a logical vector of the same length as \code{str}, \code{TRUE} for strings that are valid
genomic ranges, and \code{FALSE} otherwise.
}
\description{
Find which strings are valid genomic ranges, i.e., of the form
\{chr\}:\{start\}-\{end\} by checking: \itemize{ \item if \code{chr} is one
of the human chromosomes (1--22, X or Y); \item if \code{start} and
\code{end} are integer numbers (of at most 9 digits).}
}
\details{
Please note that in the case of the sex chromosomes, X and Y must be
uppercase, and that no check is performed to ensure that \code{start}
is less or equal than \code{end}.
}
\examples{
# Check a valid human genomic range
is_genomic_range(c("10:10-20")) # TRUE

# Test a bunch of ranges:
is_genomic_range(
  c("10:10-20", "X:1-2000", "Y:34-123")) # TRUE TRUE TRUE

# By default NAs are returned as they are.
is_genomic_range(
  c("10:10-20", "X:1-2000", NA_character_)) # TRUE TRUE NA

# Use the argument convert_NA_to_FALSE = TRUE to get FALSE instead of NA.
is_genomic_range(
  c("10:10-20", "X:1-2000", NA_character_),
  convert_NA_to_FALSE = TRUE) # TRUE TRUE FALSE

# Start and End positions must be at most a 9-digit number.
is_genomic_range(c("1:1-123456789", "1:1-0123456789")) # TRUE FALSE

# The name of the chromosome should not include 'chr'
# to be a valid genome range.
is_genomic_range("chr1:1-10") # FALSE

# No check is performed on the actual values of start and end
# to ensure that start <= end
is_genomic_range("1:10-5") # TRUE

}
