% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_links}
\alias{is_links}
\title{Does a string start with \code{_links.*}?}
\usage{
is_links(str, convert_NA_to_FALSE = FALSE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings match the the following regular expression:
\code{^_links.*$}.
}
\examples{
# Is "_links"?
is_links("_links") # TRUE

# Test a bunch of strings:
is_links(
  c("_links", "links", "A_links", "_linksB")) # TRUE FALSE FALSE TRUE

# By default NAs are returned as they are.
is_links(
  c("_links", "links", "A_links", "_linksB", NA_character_)) # TRUE FALSE FALSE TRUE NA

# Use the argument convert_NA_to_FALSE = TRUE to get FALSE instead of NA.
is_links(
  c("_links", "links", "A_links", "_linksB", NA_character_),
  convert_NA_to_FALSE = TRUE) # TRUE FALSE FALSE TRUE FALSE

}
\keyword{internal}
