% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is_paginated}
\alias{is_paginated}
\title{Is the GWAS response paginated?}
\usage{
is_paginated(content)
}
\arguments{
\item{content}{The response content as return by \code{jsonlite::fromJSON}.}
}
\value{
A logical value.
}
\description{
Checks if the response is paginated by checking if an element named 'page'
exists.
}
\keyword{internal}
