% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{is_pubmed_id}
\alias{is_pubmed_id}
\title{Is a string a PubMed ID?}
\usage{
is_pubmed_id(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid PubMed IDs (returns \code{TRUE}). PubMed IDs are
tested against the following regular expression: \code{^\\\\d+$}.
}
\keyword{internal}
