% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{is_response_successful}
\alias{is_response_successful}
\title{Was the response to a GWAS Catalog REST API request successful?}
\usage{
is_response_successful(response)
}
\arguments{
\item{response}{A response (\code{list}) as returned by \code{\link{request}}.}
}
\value{
A logical value.
}
\description{
Checks if the response code of a GWAS Catalog REST API request is
\code{200L}.
}
