% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{is_rs_id}
\alias{is_rs_id}
\title{Is a string a valid rsID?}
\usage{
is_rs_id(str, convert_NA_to_FALSE = FALSE)
}
\arguments{
\item{str}{A character vector of putative SNP reference IDs of the form
rs[0-9]+.}

\item{convert_NA_to_FALSE}{Whether to preserve \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
Returns a logical vector of the same length as \code{str},
  \code{TRUE} for strings that are valid rs IDs, and \code{FALSE} otherwise.
}
\description{
Find which strings are valid SNP reference IDs, i.e., of the form rs[0-9]+.
Please note that this only does a syntax validation on the strings. It does
not check whether the actual IDs exist in dbSNP.
}
\examples{
# Check if strings are rsIDs
is_rs_id(c("rs123")) # TRUE

# Test a vector of rsIDs
is_rs_id(
  c("rs123", "rs0001", "rs09123")) # TRUE TRUE TRUE

# By default NAs are returned as they are.
is_rs_id(
  c("rs123", "rs0001", NA_character_)) # TRUE TRUE NA

# Use the argument convert_NA_to_FALSE = TRUE to get FALSE instead of NA.
is_rs_id(
  c("rs123", "rs0001", NA_character_),
  convert_NA_to_FALSE = TRUE) # TRUE TRUE FALSE

}
