% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{is_study_accession}
\alias{is_study_accession}
\title{Is a string a GWAS Catalog study accession ID?}
\usage{
is_study_accession(str, convert_NA_to_FALSE = FALSE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid GWAS Catalog study accession IDs (returns
\code{TRUE}). Study accession IDs are tested against the following regular
expression: \code{^GCST\\\\d\{6\}$}.
}
\examples{
# Can "GCST123456" be a valid GWAS Catalog study accession ID?
is_study_accession("GCST123456") # TRUE

# Test a bunch of strings:
is_study_accession(
  c("GCST123456", "GCST000042", "123456")) # TRUE TRUE NA

# By default NAs are returned as they are.
is_study_accession(
  c("123456", "GCST123456", NA_character_)) # FALSE TRUE NA

# Use the argument convert_NA_to_FALSE = TRUE to get FALSE instead of NA.
is_study_accession(c("123456", "GCST123456", NA_character_),
                   convert_NA_to_FALSE = TRUE) # FALSE TRUE FALSE

}
