% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{is_study_id}
\alias{is_study_id}
\title{Is a string a GWAS Catalog study accession ID?}
\usage{
is_study_id(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid GWAS Catalog study accession IDs (returns
\code{TRUE}). Study accession IDs are tested against the following regular
expression: \code{^GCST\\\\d+$}.
}
\keyword{internal}
