% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-utils.R
\name{list_to_s4}
\alias{list_to_s4}
\title{Convert a named list to an S4 object}
\usage{
list_to_s4(list, class)
}
\arguments{
\item{list}{list}

\item{class}{character vector indicating the S4 class}
}
\value{
S4 object of class \code{class}.
}
\description{
Convert a named list to an S4 object
}
\keyword{internal}
