% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{loci_tbl}
\alias{loci_tbl}
\title{Creates a loci table.}
\usage{
loci_tbl(association_id = integer(), locus_id = integer(),
  haplotype_snp_count = integer(), description = character())
}
\arguments{
\item{association_id}{TODO.}

\item{locus_id}{TODO.}

\item{haplotype_snp_count}{TODO.}

\item{description}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a loci table.
}
\keyword{internal}
