% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_to_na}
\alias{missing_to_na}
\title{Recursively map all missing values to NA.}
\usage{
missing_to_na(lst)
}
\arguments{
\item{lst}{A list.}
}
\value{
A list whose missing values have been replaced by NA.
}
\description{
Recursively map all missing values to NA.
}
\details{
Note: This function is case sensitive on "NR", i.e., it will leave "nr"
untouched.
}
\keyword{internal}
