% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{n}
\alias{n}
\title{Number of GWAS Catalog Entities}
\usage{
n(x)
}
\arguments{
\item{x}{A \linkS4class{studies}, an
\linkS4class{associations}, a \linkS4class{variants}, or a
\linkS4class{tratis} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique entities in a GWAS Catalog object.
A GWAS Catalog object can be: a \linkS4class{studies}, an
\linkS4class{associations}, a \linkS4class{variants}, or a
\linkS4class{tratis} object.
}
