% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n_associations}
\alias{n_associations}
\alias{n_associations,associations-method}
\title{Number of associations}
\usage{
n_associations(x)

\S4method{n_associations}{associations}(x)
}
\arguments{
\item{x}{A \linkS4class{associations} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique associations in a \linkS4class{associations} object.
}
