% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n_studies}
\alias{n_studies}
\alias{n_studies,studies-method}
\title{Number of studies}
\usage{
n_studies(x)

\S4method{n_studies}{studies}(x)
}
\arguments{
\item{x}{A \linkS4class{studies} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique studies in a \linkS4class{studies} object.
}
