% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n_traits}
\alias{n_traits}
\alias{n_traits,traits-method}
\title{Number of traits}
\usage{
n_traits(x)

\S4method{n_traits}{traits}(x)
}
\arguments{
\item{x}{A \linkS4class{traits} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique traits in a \linkS4class{traits} object.
}
