% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n_variants}
\alias{n_variants}
\alias{n_variants,variants-method}
\title{Number of variants}
\usage{
n_variants(x)

\S4method{n_variants}{variants}(x)
}
\arguments{
\item{x}{A \linkS4class{variants} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique variants in a \linkS4class{variants} object.
}
