% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{no_content_elements}
\alias{no_content_elements}
\title{Number of elements in response of a GWAS Catalog REST API request.}
\usage{
no_content_elements(response)
}
\arguments{
\item{response}{A response (\code{list}) as returned by \code{\link{request}}.}
}
\value{
An integer value.
}
\description{
Gets the number of elements in a response (\code{list}) as returned by
\code{\link{request}}.
}
