% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nr_to_na}
\alias{nr_to_na}
\title{Recursively map all "NR" values to NA.}
\usage{
nr_to_na(lst)
}
\arguments{
\item{lst}{A list.}
}
\value{
A list whose NR elements have been replaced by NA.
}
\description{
Traverse a recursive list and map \code{"NR"} (Not Recorded) values to
\code{\link[base]{NA}}.
}
\details{
Note: This function is case sensitive on \code{"NR"}, i.e., it will leave
\code{"nr"} untouched.
}
\keyword{internal}
