% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_to_na}
\alias{null_to_na}
\title{Recursively map all NULL values to NA.}
\usage{
null_to_na(lst)
}
\arguments{
\item{lst}{A list.}
}
\value{
A list whose NULL elements have been replaced by NA.
}
\description{
Traverse a recursive list and map \code{\link[base]{NULL}} values to
\code{\link[base]{NA}}.
}
\keyword{internal}
