% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-variants.R
\name{open_in_gtex}
\alias{open_in_gtex}
\title{Browse GTEx from SNP identifiers.}
\usage{
open_in_gtex(variant_id)
}
\arguments{
\item{variant_id}{A variant identifier, a character vector.}
}
\value{
Returns \code{TRUE} if successful. Note however that this
function is run for its side effect.
}
\description{
This function launches the web browser at the GTEx Portal and opens a tab for
each SNP identifier.
}
\examples{
open_in_gtex('rs56261590')
}
