% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{peel_off_embedded}
\alias{peel_off_embedded}
\title{Peels off the _embedded tier from a JSON-list.}
\usage{
peel_off_embedded(obj)
}
\arguments{
\item{obj}{A JSON-list.}
}
\value{
A JSON-list.
}
\description{
This function removes an element named \code{_embedded} from the element
\code{content} moving all of its contents one level up. If
\code{obj$content$`_embedded`} does not exist then returns \code{obj}
untouched.
}
\keyword{internal}
