% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{platforms_tbl}
\alias{platforms_tbl}
\title{Creates a platforms table.}
\usage{
platforms_tbl(study_id = character(), manufacturer = character())
}
\arguments{
\item{study_id}{GWAS Catalog study accession identifier.}

\item{manufacturer}{Platform manufacturer.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a platforms table.
}
\keyword{internal}
