% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{publication_to_tibble}
\alias{publication_to_tibble}
\title{Convert a publication list object to a tibble.}
\usage{
publication_to_tibble(publication_content)
}
\arguments{
\item{publication_content}{The publication content (list).}
}
\value{
A \code{\link[tibble]{tibble}} version of \code{publication_content}.
}
\description{
Convert a publication list object to a tibble.
}
\keyword{internal}
