% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{reported_genes_tbl}
\alias{reported_genes_tbl}
\title{Creates an authors' reported genes table.}
\usage{
reported_genes_tbl(association_id = integer(), locus_id = integer(),
  gene_name = character())
}
\arguments{
\item{association_id}{TODO.}

\item{locus_id}{TODO.}

\item{gene_name}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an authors' reported genes table.
}
\keyword{internal}
