% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request a GWAS REST API resource}
\usage{
request(resource_url = "/", base_url = gwas_rest_api_base_url,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{resource_url}{Endpoint URL. The endpoint is internally appended to the
\code{base_url}.}

\item{base_url}{The GWAS REST API base URL (one should not need to change its
default value).}

\item{verbose}{Whether to be chatty.}

\item{warnings}{Whether to print warnings.}
}
\value{
A list of two elements: (i) response code, and (ii) the parsed JSON
  content.
}
\description{
Performs a GET request on the specified \code{resource_url}.
}
