% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{risk_alleles_tbl}
\alias{risk_alleles_tbl}
\title{Creates a risk alleles table.}
\usage{
risk_alleles_tbl(association_id = integer(), locus_id = integer(),
  variant_id = character(), risk_allele = character(),
  risk_frequency = double(), genome_wide = logical(),
  limited_list = logical())
}
\arguments{
\item{association_id}{TODO.}

\item{locus_id}{TODO.}

\item{variant_id}{TODO.}

\item{risk_allele}{TODO.}

\item{risk_frequency}{TODO.}

\item{genome_wide}{TODO.}

\item{limited_list}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a risk alleles table.
}
\keyword{internal}
