% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-utils.R
\name{s4_to_list}
\alias{s4_to_list}
\title{Convert an S4 object into a list}
\usage{
s4_to_list(s4_obj)
}
\arguments{
\item{s4_obj}{an S4 object}
}
\value{
A list version of the S4 object.
}
\description{
Convert an S4 object into a list
}
\keyword{internal}
