% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{skip_if_testing_is_fast}
\alias{skip_if_testing_is_fast}
\title{Skips a test if TEST_FAST is 'true'}
\usage{
skip_if_testing_is_fast()
}
\description{
This function checks the value of \code{TEST_FAST}, if it is 'true' it skips
the test otherwise it still runs the following expectations.
Please note that this function is to be used in gwasrapidd development.
}
\keyword{internal}
