% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{snp_content_to_tibble}
\alias{snp_content_to_tibble}
\title{Convert a SNP list object to a tibble.}
\usage{
snp_content_to_tibble(snp_content)
}
\arguments{
\item{snp_content}{The response content (list).}
}
\value{
A \code{\link[tibble]{tibble}} where each row pertains a SNP, with
  the following columns: \itemize{ \item \code{rsId}, a \code{character}
  vector. \item \code{merged}, an \code{integer} vector. \item
  \code{chromosomeName}, a \code{character} vector. \item
  \code{chromosomePosition}, an \code{integer} vector. \item
  \code{region.name}, a \code{character} vector. \item
  \code{functionalClass}, a \code{character} vector. \item
  \code{lastUpdateDate}, a \code{character} vector. \item
  \code{genomicContexts}, a \code{list}. }
}
\description{
Converts the content returned by a request to either a
\code{"/snpLocation/{chr:start-end}"} or a
\code{"singleNucleotidePolymorphisms/{rsId}"} to a
\code{\link[tibble]{tibble}}.
}
\keyword{internal}
