% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{studies}
\alias{studies}
\title{Constructor for the S4 studies object.}
\usage{
studies(studies = studies_tbl(),
  genotyping_techs = genotyping_techs_tbl(),
  platforms = platforms_tbl(), ancestries = ancestries_tbl(),
  ancestral_groups = ancestral_groups_tbl(),
  countries_of_origin = countries_tbl(),
  countries_of_recruitment = countries_tbl(),
  publications = publications_tbl())
}
\arguments{
\item{studies}{TODO.}

\item{genotyping_techs}{TODO.}

\item{platforms}{TODO.}

\item{ancestries}{TODO.}

\item{ancestral_groups}{TODO.}

\item{countries_of_origin}{TODO.}

\item{countries_of_recruitment}{TODO.}

\item{publications}{TODO.}
}
\value{
An object of class \linkS4class{studies}.
}
\description{
Constructor for the S4 \linkS4class{studies} object.
}
\keyword{internal}
