% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{studies_drop_na}
\alias{studies_drop_na}
\title{Drop any NA studies.}
\usage{
studies_drop_na(s4_studies)
}
\arguments{
\item{s4_studies}{An object of class \linkS4class{studies}.}
}
\value{
An object of class \linkS4class{studies}.
}
\description{
This function takes a studies S4 object and removes any study identifiers
that might have been NA. This ensures that there is always a non-NA
\code{study_id} value in all tables. This is important as the \code{study_id}
is the primary key.
}
\keyword{internal}
