% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_content_to_tibble}
\alias{study_content_to_tibble}
\title{Convert a study list object to a tibble.}
\usage{
study_content_to_tibble(study_content)
}
\arguments{
\item{study_content}{The response content (list).}
}
\value{
TODO.
}
\description{
Converts the content returned by a request to
\code{"/studies/{study_accession}"} to a
\code{\link[tibble]{tibble}}.
}
\keyword{internal}
