% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subset-associations}
\alias{subset-associations}
\alias{[,associations,missing,missing,missing-method}
\alias{[,associations,numeric,missing,missing-method}
\alias{[,associations,character,missing,missing-method}
\title{Subset an associations object}
\usage{
\S4method{[}{associations,missing,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{associations,numeric,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{associations,character,missing,missing}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \linkS4class{associations} object.}

\item{i}{Position of the identifier or the name of the identifier itself.}

\item{j}{Not used.}

\item{...}{Additional arguments not used here.}

\item{drop}{Not used.}
}
\value{
A \linkS4class{associations} object.
}
\description{
You can subset \linkS4class{associations} by identifier or by position using the
\code{`[`} operator.
}
\examples{
# Subset an associations object by identifier
associations_ex01['22505']

# Or by its position in table associations
associations_ex01[2]

# Keep all associations except the second
associations_ex01[-2]

}
