% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,studies,missing,missing,missing-method}
\alias{[,studies,numeric,missing,missing-method}
\alias{[,studies,character,missing,missing-method}
\alias{[,associations,missing,missing,missing-method}
\alias{[,associations,numeric,missing,missing-method}
\alias{[,associations,character,missing,missing-method}
\alias{[,variants,missing,missing,missing-method}
\alias{[,variants,numeric,missing,missing-method}
\alias{[,variants,character,missing,missing-method}
\alias{[,traits,missing,missing,missing-method}
\alias{[,traits,numeric,missing,missing-method}
\alias{[,traits,character,missing,missing-method}
\title{Subsetting GWAS Catalog objects}
\usage{
\S4method{[}{studies,missing,missing,missing}(x, i)

\S4method{[}{studies,numeric,missing,missing}(x, i)

\S4method{[}{studies,character,missing,missing}(x, i)

\S4method{[}{associations,missing,missing,missing}(x, i)

\S4method{[}{associations,numeric,missing,missing}(x, i)

\S4method{[}{associations,character,missing,missing}(x, i)

\S4method{[}{variants,missing,missing,missing}(x, i)

\S4method{[}{variants,numeric,missing,missing}(x, i)

\S4method{[}{variants,character,missing,missing}(x, i)

\S4method{[}{traits,missing,missing,missing}(x, i)

\S4method{[}{traits,numeric,missing,missing}(x, i)

\S4method{[}{traits,character,missing,missing}(x, i)
}
\arguments{
\item{x}{A \linkS4class{studies}, \linkS4class{associations},
\linkS4class{variants}, or \linkS4class{traits} object.}

\item{i}{Position of the identifier or the name of the identifier itself.}
}
\value{
An object of the same class as \code{x}, i.e.,
  \linkS4class{studies}, \linkS4class{associations}, \linkS4class{variants},
  or \linkS4class{traits}.
}
\description{
This function allows you to subset \linkS4class{studies},
\linkS4class{associations}, \linkS4class{variants}, or \linkS4class{traits},
by their respective identifiers or by position.
}
