% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sure.R
\name{sure}
\alias{sure}
\title{Are you sure?}
\usage{
sure(
  before_question = NULL,
  after_saying_no = NULL,
  after_saying_yes = NULL,
  default_answer = NULL
)
}
\arguments{
\item{before_question}{String with message to be printed before question.}

\item{after_saying_no}{String with message to be printed after answering
\code{'no'}.}

\item{after_saying_yes}{String with message to be printed after answering
\code{'yes'}.}

\item{default_answer}{String with answer to question, if run in
non-interactive mode.}
}
\value{
A logical indicating if answer was \code{'yes'}/\code{'y'}
  (\code{TRUE}) or otherwise (\code{FALSE}).
}
\description{
This function asks you interactively for permission to continue or not. You
can specify a custom message before the question and also different messages
for a both a positive and negative answer.
}
\details{
If you run this function in non-interactive mode, you should pass an
automatic answer to \code{default_answer}: \code{'yes'} or \code{'no'}.
}
\keyword{internal}
