% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{trait_to_variant}
\alias{trait_to_variant}
\title{Map an EFO trait id to a variant id}
\usage{
trait_to_variant(efo_id, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{efo_id}{A character vector of EFO trait identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}
}
\value{
A dataframe of two identifiers. First column is the EFO trait
  identifier and the second column is the variant identifier.
}
\description{
Map an EFO trait id to a variant identifier.
}
\examples{
# Map EFO trait identifiers to variant identifiers
trait_to_variant('EFO_0005229')

}
