% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\docType{class}
\name{traits-class}
\alias{traits-class}
\title{An S4 class to represent a set of GWAS Catalog EFO traits.}
\description{
The traits object consists of one slot only, a table
(\code{\link[tibble]{tibble}}) of GWAS Catalog EFO traits. Each EFO trait is
an observation (row) in the \code{traits} table --- main table.
}
\section{Slots}{

\describe{
\item{\code{traits}}{A \code{\link[tibble]{tibble}} listing EFO traits. Columns:
\describe{
\item{efo_id}{\href{https://www.ebi.ac.uk/efo/}{EFO} identifier.}
\item{trait}{\href{https://www.ebi.ac.uk/efo/}{EFO} trait description.}
\item{uri}{The full URI of the \href{https://www.ebi.ac.uk/efo/}{EFO} term.}
}}
}}

