% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\name{traits}
\alias{traits}
\title{Constructor for the S4 traits object.}
\usage{
traits(traits = traits_tbl())
}
\arguments{
\item{traits}{A \code{\link{traits_tbl}} tibble.}
}
\value{
An object of class \linkS4class{traits}.
}
\description{
Constructor for the S4 \linkS4class{traits} object.
}
\keyword{internal}
