% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\name{traits_drop_na}
\alias{traits_drop_na}
\title{Drop any NA traits.}
\usage{
traits_drop_na(s4_traits)
}
\arguments{
\item{s4_traits}{An object of class \linkS4class{traits}.}
}
\value{
An object of class \linkS4class{traits}.
}
\description{
This function takes a traits S4 object and removes any EFO trait identifiers
that might have been NA. This ensures that there is always a non-NA
\code{efo_id} value in all tables. This is important as the \code{efo_id}
is the primary key.
}
\keyword{internal}
