% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\name{traits_tbl}
\alias{traits_tbl}
\title{Creates a traits tibble}
\usage{
traits_tbl(efo_id = character(), trait = character(), uri = character())
}
\arguments{
\item{efo_id}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} identifiers.}

\item{trait}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} trait descriptions.}

\item{uri}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} URIs.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
to the function.
}
\description{
Creates a traits tibble.
}
\keyword{internal}
