% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{tws}
\alias{tws}
\title{Trim whitespace.}
\usage{
tws(x)
}
\arguments{
\item{x}{A \code{\link[base]{character}} vector.}
}
\value{
A character vector.
}
\description{
This function removes leading and trailing white space from strings. Note:
this function does no checking on input for performance reasons. So make sure
the input is really a character vector.
}
\keyword{internal}
