% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{v_entrez_ids_tbl}
\alias{v_entrez_ids_tbl}
\title{Creates a gene Entrez identifiers table.}
\usage{
v_entrez_ids_tbl(variant_id = character(), gene_name = character(),
  entrez_id = character())
}
\arguments{
\item{variant_id}{TODO.}

\item{gene_name}{TODO.}

\item{entrez_id}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a gene Entrez identifiers table.
}
\keyword{internal}
