% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-utils.R
\name{variant_name}
\alias{variant_name}
\title{Extract variant identifiers from strings of the form rs123-G}
\usage{
variant_name(risk_allele_names)
}
\arguments{
\item{risk_allele_names}{}
}
\value{
A character vector of variant names.
}
\description{
This function parses strings of the form \code{"rs123-G"} and returns
the name of the variant; it uses the regex \code{-([ATCG\\?]+)?$}.
}
\keyword{internal}
