% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{variant_to_association}
\alias{variant_to_association}
\title{Map a variant id to an association id}
\usage{
variant_to_association(variant_id, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{variant_id}{A character vector of variant identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}
}
\value{
A dataframe of two identifiers. First column is the variant
  identifier and the second column is the association identifier.
}
\description{
Map a variant identifier to an association identifier.
}
\examples{
# Map GWAS variant identifiers to association identifiers
variant_to_association(c('rs7904579', 'rs138331350'))

}
