% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{variant_to_study}
\alias{variant_to_study}
\title{Map a variant id to a study id}
\usage{
variant_to_study(variant_id, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{variant_id}{A character vector of variant identifiers.}

\item{verbose}{Whether the function should be
verbose about the different queries or not.}

\item{warnings}{Whether to print warnings.}
}
\value{
A dataframe of two identifiers. First column is the variant
  identifier and the second column is the study identifier.
}
\description{
Map a variant identifier to a study accession identifier.
}
\examples{
\dontrun{
# Map GWAS variant identifiers to study identifiers
variant_to_study(c('rs7904579', 'rs138331350'))
}

}
