% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\docType{class}
\name{variants-class}
\alias{variants-class}
\title{An S4 class to represent a set of GWAS Catalog variants}
\description{
The variants object consists of four slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of GWAS Catalog variants Each variant is an observation (row) in the
variants table --- main table. All tables have the column \code{variant_id} as
primary key.
}
\section{Slots}{

\describe{
\item{\code{variants}}{A \code{\link[tibble]{tibble}} listing variants. Columns:
\describe{
\item{TODO}{TODO}
}}

\item{\code{genomic_contexts}}{A \code{\link[tibble]{tibble}} listing genomic
  contexts associated with each variant. Columns:
\describe{
\item{TODO}{TODO}
}}

\item{\code{ensembl_ids}}{A \code{\link[tibble]{tibble}} listing gene Ensembl
  identifiers associated with each genomic context. Columns:
\describe{
\item{TODO}{TODO}
}}

\item{\code{entrez_ids}}{A \code{\link[tibble]{tibble}} listing gene Entrez
  identifiers associated with each genomic context. Columns:
\describe{
\item{TODO}{TODO}
}}
}}

