% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{variants_drop_na}
\alias{variants_drop_na}
\title{Drop any NA variants.}
\usage{
variants_drop_na(s4_variants)
}
\arguments{
\item{s4_variants}{An object of class \linkS4class{variants}.}
}
\value{
An object of class \linkS4class{variants}.
}
\description{
This function takes a variants S4 object and removes any variant identifiers
that might have been NA. This ensures that there is always a non-NA
\code{variant_id} value in all tables. This is important as the
\code{variant_id} is the primary key.
}
\keyword{internal}
