% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{variants_tbl}
\alias{variants_tbl}
\title{Creates a variants table.}
\usage{
variants_tbl(variant_id = character(), merged = integer(),
  functional_class = character(), chromosome_name = character(),
  chromosome_position = integer(), chromosome_region = character(),
  last_update_date = lubridate::ymd_hms())
}
\arguments{
\item{variant_id}{TODO.}

\item{merged}{TODO.}

\item{functional_class}{TODO.}

\item{chromosome_name}{TODO.}

\item{chromosome_position}{TODO.}

\item{chromosome_region}{TODO.}

\item{last_update_date}{TODO.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a variants table.
}
\keyword{internal}
