% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{associations}
\alias{associations}
\title{Constructor for the S4 associations object.}
\usage{
associations(
  associations = associations_tbl(),
  loci = loci_tbl(),
  risk_alleles = risk_alleles_tbl(),
  genes = reported_genes_tbl(),
  ensembl_ids = ensembl_ids_tbl(),
  entrez_ids = entrez_ids_tbl()
)
}
\arguments{
\item{associations}{An \code{\link{associations_tbl}} tibble.}

\item{loci}{A \code{\link{loci_tbl}} tibble.}

\item{risk_alleles}{A \code{\link{risk_alleles_tbl}} tibble.}

\item{genes}{A \code{\link{reported_genes_tbl}} tibble.}

\item{ensembl_ids}{A \code{\link{ensembl_ids_tbl}} tibble.}

\item{entrez_ids}{A \code{\link{entrez_ids_tbl}} tibble.}
}
\value{
An object of class \linkS4class{associations}.
}
\description{
Constructor for the S4 \linkS4class{associations} object.
}
\keyword{internal}
