% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{entrez_ids_tbl}
\alias{entrez_ids_tbl}
\title{Creates an Entrez gene identifiers' table.}
\usage{
entrez_ids_tbl(
  association_id = character(),
  locus_id = integer(),
  gene_name = character(),
  entrez_id = character()
)
}
\arguments{
\item{association_id}{A character vector of association identifiers.}

\item{locus_id}{An integer vector of locus identifiers.}

\item{gene_name}{A character vector of gene symbol according to
\href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}

\item{entrez_id}{A character vector of Entrez identifiers.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an Entrez gene identifiers' table.
}
\keyword{internal}
