% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{filter_by_id}
\alias{filter_by_id}
\alias{filter_by_id,studies,character-method}
\alias{filter_by_id,associations,integer-method}
\alias{filter_by_id,variants,character-method}
\alias{filter_by_id,traits,character-method}
\title{Filter GWAS Catalog objects by Id}
\usage{
filter_by_id(x, id)

\S4method{filter_by_id}{studies,character}(x, id)

\S4method{filter_by_id}{associations,integer}(x, id)

\S4method{filter_by_id}{variants,character}(x, id)

\S4method{filter_by_id}{traits,character}(x, id)
}
\arguments{
\item{x}{An object of class either \linkS4class{studies},
\linkS4class{associations}, \linkS4class{variants}, or
\linkS4class{traits}}

\item{id}{Id, e.g., \code{"GCST001374"} in the case of \linkS4class{studies},
or an integer \code{24299710} if an association, or \code{"rs10910092"} if a
variant, or \code{"EFO_0004761"}.}
}
\description{
Use \code{filter_by_id} to filter GWAS Catalog objects by their respective
\code{id}.
}
\keyword{internal}
