% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Filter GWAS Catalog objects by identifier.}
\usage{
filter_by_id(x, id)
}
\arguments{
\item{x}{An object of class either \linkS4class{studies},
\linkS4class{associations}, \linkS4class{variants}, or
\linkS4class{traits}.}

\item{id}{Identifier.}
}
\value{
Returns an object of class either \linkS4class{studies},
  \linkS4class{associations}, \linkS4class{variants}, or
  \linkS4class{traits}.
}
\description{
Use \code{filter_by_id} to filter GWAS Catalog objects by their respective
identifier (\code{id}).
}
\keyword{internal}
