% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{loci_tbl}
\alias{loci_tbl}
\title{Creates a loci table.}
\usage{
loci_tbl(
  association_id = character(),
  locus_id = integer(),
  haplotype_snp_count = integer(),
  description = character()
)
}
\arguments{
\item{association_id}{A character vector of association identifiers.}

\item{locus_id}{An integer vector of locus identifiers.}

\item{haplotype_snp_count}{An integer vector indicating the number of
variants in the haplotype.}

\item{description}{A character vector of descriptions, one per locus
identifier.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a loci table.
}
\keyword{internal}
