% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{missing_to_na}
\alias{missing_to_na}
\title{Recursively map all missing values to \code{NA}.}
\usage{
missing_to_na(lst, na = NA_character_)
}
\arguments{
\item{lst}{A list.}
}
\value{
A list whose missing values have been replaced by NA.
}
\description{
Traverse a recursive list and map missing values to \code{\link[base]{NA}}.
Missing values are: \itemize{ \item \code{NULL} \item \code{"NR"} }
}
\details{
Note: This function is case sensitive on \code{"NR"}, i.e., it will leave
\code{"nr"} untouched.
}
\keyword{internal}
