% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n}
\alias{n}
\alias{n,studies-method}
\alias{n,associations-method}
\alias{n,variants-method}
\alias{n,traits-method}
\title{Number of GWAS Catalog entities}
\usage{
n(x)

\S4method{n}{studies}(x)

\S4method{n}{associations}(x)

\S4method{n}{variants}(x)

\S4method{n}{traits}(x)
}
\arguments{
\item{x}{A \linkS4class{studies}, an
\linkS4class{associations}, a \linkS4class{variants}, or a
\linkS4class{traits} object.}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique entities in a GWAS Catalog object.
}
