% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{n_studies}
\alias{n_studies}
\alias{n_studies,studies-method}
\title{#' @export
setMethod("[",
          signature(x = "studies", i = "missing", j = "missing", drop = "missing"),
          definition = function(x) x)}
\usage{
n_studies(x)

\S4method{n_studies}{studies}(x)
}
\arguments{
\item{x}{A \linkS4class{studies} object.}
}
\value{
An integer scalar.
}
\description{
#' @export
setMethod("[",
          signature(x = "studies", i = "numeric", j = "missing", drop = "missing"),
          definition = function(x, i) {
            study_ids <- unique(x@studies$study_id)[i]
            filter_by_id(x, id = study_ids)
          })
}
\details{
#' @export
setMethod("[",
          signature(x = "studies", i = "character", j = "missing", drop = "missing"),
          definition = function(x, i) {
            filter_by_id(x, id = i)
          })
Number of studies

This function returns the number of unique studies in a \linkS4class{studies} object.
}
