% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{reported_genes_tbl}
\alias{reported_genes_tbl}
\title{Creates an authors' reported genes table.}
\usage{
reported_genes_tbl(
  association_id = character(),
  locus_id = integer(),
  gene_name = character()
)
}
\arguments{
\item{association_id}{A character vector of association identifiers.}

\item{locus_id}{An integer vector of locus identifiers.}

\item{gene_name}{A character vector of gene symbol according to
\href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates an authors' reported genes table.
}
\keyword{internal}
