% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-associations.R
\name{risk_alleles_tbl}
\alias{risk_alleles_tbl}
\title{Creates a risk alleles table.}
\usage{
risk_alleles_tbl(
  association_id = character(),
  locus_id = integer(),
  variant_id = character(),
  risk_allele = character(),
  risk_frequency = double(),
  genome_wide = logical(),
  limited_list = logical()
)
}
\arguments{
\item{association_id}{A character vector of association identifiers.}

\item{locus_id}{An integer vector of locus identifiers.}

\item{variant_id}{A character vector of variant identifiers.}

\item{risk_allele}{A character vector of risk or effect allele names.}

\item{risk_frequency}{A numeric vector of the frequency of risk or effect
alleles.}

\item{genome_wide}{A logical vector.}

\item{limited_list}{A logical vector.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a risk alleles table.
}
\keyword{internal}
