% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{skip_if_testing_is_fast}
\alias{skip_if_testing_is_fast}
\title{Skips a test if TEST_FAST is 'true'}
\usage{
skip_if_testing_is_fast()
}
\value{
Returns \code{TRUE} if the environment variable \code{'TEST_FAST'} is
  set to \code{'false'}, or does not return anything but triggers the side
  effect of skipping the next test (with testthat).
}
\description{
This function checks the value of \code{TEST_FAST}, if it is 'true' it skips
the test otherwise it still runs the following expectations. Please note that
this function is to be used in gwasrapidd development.
}
\keyword{internal}
