% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-studies.R
\name{studies}
\alias{studies}
\title{Constructor for the S4 studies object.}
\usage{
studies(
  studies = studies_tbl(),
  genotyping_techs = genotyping_techs_tbl(),
  platforms = platforms_tbl(),
  ancestries = ancestries_tbl(),
  ancestral_groups = ancestral_groups_tbl(),
  countries_of_origin = countries_tbl(),
  countries_of_recruitment = countries_tbl(),
  publications = publications_tbl()
)
}
\arguments{
\item{studies}{A \code{\link{studies_tbl}} tibble.}

\item{genotyping_techs}{A \code{\link{genotyping_techs_tbl}} tibble.}

\item{platforms}{A \code{\link{platforms_tbl}} tibble.}

\item{ancestries}{A \code{\link{ancestries_tbl}} tibble.}

\item{ancestral_groups}{A \code{\link{ancestral_groups_tbl}} tibble.}

\item{countries_of_origin}{A \code{\link{countries_tbl}} tibble.}

\item{countries_of_recruitment}{A \code{\link{countries_tbl}} tibble.}

\item{publications}{A \code{\link{publications_tbl}} tibble.}
}
\value{
An object of class \linkS4class{studies}.
}
\description{
Constructor for the S4 \linkS4class{studies} object.
}
\keyword{internal}
