% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_content_to_tibble}
\alias{study_content_to_tibble}
\title{Convert a study list object to a tibble.}
\usage{
study_content_to_tibble(study_content)
}
\arguments{
\item{study_content}{The study content (list).}
}
\value{
A \code{\link[tibble]{tibble}} version of \code{study_content}.
}
\description{
Convert a study list object to a tibble.
}
\keyword{internal}
