% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\name{traits}
\alias{traits}
\title{Constructor for the S4 traits object.}
\usage{
traits(traits = traits_tbl())
}
\arguments{
\item{efo_id}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} identifiers.}

\item{trait}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} trait descriptions.}

\item{uri}{A \code{character} vector of
\href{https://www.ebi.ac.uk/efo/}{EFO} URIs.}
}
\value{
An object of class \linkS4class{traits}.
}
\description{
Constructor for the S4 \linkS4class{traits} object.
}
\keyword{internal}
