% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{v_entrez_ids_tbl}
\alias{v_entrez_ids_tbl}
\title{Creates a gene Entrez identifiers table.}
\usage{
v_entrez_ids_tbl(
  variant_id = character(),
  gene_name = character(),
  entrez_id = character()
)
}
\arguments{
\item{variant_id}{A character vector of variant identifiers.}

\item{gene_name}{A character vector of gene symbols according to
\href{https://www.genenames.org/}{HUGO Gene Nomenclature (HGNC)}.}

\item{entrez_id}{A character vector of Entrez identifiers.}
}
\value{
A \code{\link[tibble]{tibble}} whose columns are the named arguments
  to the function.
}
\description{
Creates a gene Entrez identifiers table.
}
\keyword{internal}
