% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variants.R
\name{variants}
\alias{variants}
\title{Constructor for the S4 variants object.}
\usage{
variants(
  variants = variants_tbl(),
  genomic_contexts = genomic_contexts_tbl(),
  ensembl_ids = v_ensembl_ids_tbl(),
  entrez_ids = v_entrez_ids_tbl()
)
}
\arguments{
\item{variants}{A \code{\link{variants_tbl}} tibble.}

\item{genomic_contexts}{A \code{\link{genomic_contexts_tbl}} tibble.}

\item{ensembl_ids}{A \code{\link{v_ensembl_ids_tbl}} tibble.}

\item{entrez_ids}{A \code{\link{v_entrez_ids_tbl}} tibble.}
}
\value{
An object of class \linkS4class{variants}.
}
\description{
Constructor for the S4 \linkS4class{variants} object.
}
\keyword{internal}
